Move character into own component
